requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
numpy==1.26.4
scipy==1.11.4
spacy==3.7.4
feedparser==6.0.11
newspaper3k==0.2.8
readability-lxml==0.8.1
beautifulsoup4==4.12.3
requests==2.32.3
gensim==4.3.2
nltk==3.8.1

